Variable Explorer: Rename dicteditorutils to varexp/utils
