+++ spyderlib/widgets/varexp/utils.py
